RDF graph isomorphism comparator
Implement a graph isomorphism function. This checks whether two RDF
graphs are isomorphic to each other. This would be trivial, if it
weren't for the bnodes. Having bnodes makes it much harder, and
ultimately devolves to trying every combination of matching a bnode
in one graph against a bnode in the other. However, we try not to
do that, and use information about what we can recognise, that each
bnode is connected to, to guess likely identities for the bnodes.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>