Fix bug: don't automatically ground every ungrounded pair
If we have precisely one node in each graph that shares a signature,
we can't simply assume that they're isomorphic. Instead, we need to do
the whole trial-and-error depth-first search process on every such pair
in turn.

So, remove the automatic identification of paired nodes, and just head
straight into the pairing search -- and try all possible pairs from all
possible groups, not just the first group.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>