Rewrite ntriples parser to allow for incremental parsing
This is a very heavy rewrite of the ntriples parser, so that it can parse one
triple at a time while maintaining parse state. This allows us to implement
an incremental parser, which will return from parsing with a batch of triples,
rather than attempting to read everything into RAM at once.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>