Bnodes: generate globally-unique bnode IDs
When we generate bnodes, we can't guarantee that the store
representation in erlang code will be able to maintain uniqueness of
the bnode IDs across a restart of the system, for example (say, a
counter being reset to the same value every time).

So, let's instead generate a probably-unique ID each time. We do this
with a bnode-id server, which initialises with two large random numbers
(a serial and a suffix), incrementing the serial each time a new bnode
is generated, and concatenating them. When the serial counter is
exhausted, generate a new suffix and reset the counter.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>