Test suite generation: Always create two stores
The Turtle test suite has a mix of tests: some one-store tests (parser
behaviour), and some two-store tests (ensuring correctness of the resulting
triples). The current code can't cope with that. Fix it by dropping the
parameter for the number of stores to create, and always creating two
stores, even if only one is used.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>