NTriples parser: parse XSD typed elements as their erlang datatype
Teach the NTriples parser a bit about XSD datatypes, so that it can
convert them correctly into an erlang representation.

Note that the test for isomorphism of datatypes (easyrdf-bad-06) seems
to be correct (the two graphs are not isomorphic) only where the XSD
datatypes are *not* in known (not in the L2V mapping). Where they are
known, the graph is identical. So, also remove that test from the test
suite.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>