Change auto-generated test suites to use actual test name in function
Instead of transforming the test name to make it a valid erlang function
name, quote it with '' to make it an explicit atom, regardless of its
content.

This makes no difference to the execution of the tests, but does mean that
we can carry exact metadata through the process, for generating test
reports.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>