Make the triple and quad constructors more flexible and easier to use
The constructors for triples and quads were hard to use, because they
required the use of new_literal/1,2, new_iri/1, and new_bnode/0 to build
the components of the triple. Particularly for literals and IRIs, which
are the most common cases, this gets painful very quickly.

Extend the API to turn binaries into IRIs, and other things into literals.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>